refactor: move logging.py to "mod" directory
